+++ src/v1/function-configuration.ts
@@ -182,7 +182,7 @@ export interface RuntimeOptions {
   /**
    * Amount of memory to allocate to the function.
    */
-  memory?: typeof VALID_MEMORY_OPTIONS[number] | Expression<number> | ResetValue;
+  memory?: (typeof VALID_MEMORY_OPTIONS)[number] | Expression<number> | ResetValue;
   /**
    * Timeout for the function in seconds, possible values are 0 to 540.
    */
@@ -210,17 +210,17 @@ export interface RuntimeOptions {
   /**
    * Egress settings for VPC connector.
    */
-  vpcConnectorEgressSettings?: typeof VPC_EGRESS_SETTINGS_OPTIONS[number] | ResetValue;
+  vpcConnectorEgressSettings?: (typeof VPC_EGRESS_SETTINGS_OPTIONS)[number] | ResetValue;
 
   /**
    * Specific service account for the function to run as.
    */
-  serviceAccount?: "default" | string | ResetValue;
+  serviceAccount?: "default" | string | Expression<string> | ResetValue;
 
   /**
    * Ingress settings which control where this function can be called from.
    */
-  ingressSettings?: typeof INGRESS_SETTINGS_OPTIONS[number] | ResetValue;
+  ingressSettings?: (typeof INGRESS_SETTINGS_OPTIONS)[number] | ResetValue;
 
   /**
    * User labels to set on the function.
@@ -270,7 +270,7 @@ export interface DeploymentOptions extends RuntimeOptions {
   /**
    * Regions where function should be deployed.
    */
-  regions?: Array<typeof SUPPORTED_REGIONS[number] | string | Expression<string> | ResetValue>;
+  regions?: Array<(typeof SUPPORTED_REGIONS)[number] | string | Expression<string> | ResetValue>;
   /**
    * Schedule for the scheduled function.
    */