+++ src/v2/providers/alerts/crashlytics.ts
@@ -255,7 +255,7 @@ export interface CrashlyticsOptions extends options.EventHandlerOptions {
   /**
    * Specific service account for the function to run as.
    */
-  serviceAccount?: string | ResetValue;
+  serviceAccount?: string | Expression<string> | ResetValue;
 
   /**
    * Ingress settings which control where this function can be called from.