+++ CHANGELOG.md
@@ -1,1 +1,3 @@
-- Adds support for region params.
+- Enhance firstore triggers (#1358).
+- Allow parametrized string type in ServiceAccount fields in Functions and trigger configs (#1309)
+- Adds support for region params (#1353).